requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.35
python-dotenv>=1.0.1
google-genai>=1.0.0
openai>=1.54.0
yt-dlp>=2024.10.22
youtube-transcript-api>=0.6.3
httpx>=0.27.2
jinja2>=3.1.4
python-multipart>=0.0.12
google-api-python-client>=2.149.0
